Add files via upload
In this coding session, participants will dive into the world of Python programming by creating a Tip Calculator. The project prompts users to input the total bill amount and the percentage of tip they wish to leave. The program then calculates the tip amount and displays the total bill, making it a hands-on exercise in basic user interaction and mathematical operations. This engaging project is a stepping stone in applying Python skills to real-world scenarios, setting the stage for further exploration and learning throughout the "100 Days of Code" challenge.